Remove multi warning when traindata not present
