+++ src/lib_ccx/608_smptett.c
@@ -106,30 +106,12 @@ void write_stringz_as_smptett(char *string, struct encoder_ctx *context, LLONG m
 
 int write_cc_bitmap_as_smptett(struct cc_subtitle *sub, struct encoder_ctx *context)
 {
-	struct spupng_t *sp = (struct spupng_t *)context->out->spupng_data;
-	int x_pos, y_pos, width, height, i;
-	int x, y, y_off, x_off, ret;
-	uint8_t *pbuf;
-	//char *filename;
+	int ret = 0;
 	struct cc_bitmap* rect;
-	png_color *palette = NULL;
-	png_byte *alpha = NULL;
-#ifdef ENABLE_OCR
-	char*str = NULL;
-#endif
-	//int used;
-#ifdef ENABLE_OCR
-	unsigned h1, m1, s1, ms1;
-	unsigned h2, m2, s2, ms2;
-#endif
 	LLONG ms_start, ms_end;
 	//char timeline[128];
 	int len = 0;
 
-        x_pos = -1;
-        y_pos = -1;
-        width = 0;
-        height = 0;
 
 	if (context->prev_start != -1 && (sub->flags & SUB_EOD_MARKER))
 	{
@@ -150,96 +132,34 @@ int write_cc_bitmap_as_smptett(struct cc_subtitle *sub, struct encoder_ctx *cont
 	if(sub->nb_data == 0 )
 		return 0;
 	rect = sub->data;
-	for(i = 0;i < sub->nb_data;i++)
-	{
-		if(x_pos == -1)
-		{
-			x_pos = rect[i].x;
-			y_pos = rect[i].y;
-			width = rect[i].w;
-			height = rect[i].h;
-		}
-		else
-		{
-			if(x_pos > rect[i].x)
-			{
-				width += (x_pos - rect[i].x);
-				x_pos = rect[i].x;
-			}
 
-                        if (rect[i].y < y_pos)
-                        {
-                                height += (y_pos - rect[i].y);
-                                y_pos = rect[i].y;
-                        }
-
-                        if (rect[i].x + rect[i].w > x_pos + width)
-                        {
-                                width = rect[i].x + rect[i].w - x_pos;
-                        }
-
-                        if (rect[i].y + rect[i].h > y_pos + height)
-                        {
-                                height = rect[i].y + rect[i].h - y_pos;
-                        }
-
-		}
-	}
 	if ( sub->flags & SUB_EOD_MARKER )
 		context->prev_start =  sub->start_time;
-	pbuf = (uint8_t*) malloc(width * height);
-	memset(pbuf, 0x0, width * height);
 
-	for(i = 0;i < sub->nb_data;i++)
-	{
-		x_off = rect[i].x - x_pos;
-		y_off = rect[i].y - y_pos;
-		for (y = 0; y < rect[i].h; y++)
-		{
-			for (x = 0; x < rect[i].w; x++)
-				pbuf[((y + y_off) * width) + x_off + x] = rect[i].data[0][y * rect[i].w + x];
-
-		}
-	}
-	palette = (png_color*) malloc(rect[0].nb_colors * sizeof(png_color));
-	if(!palette)
-	{
-		ret = -1;
-		goto end;
-	}
-        alpha = (png_byte*) malloc(rect[0].nb_colors * sizeof(png_byte));
-        if(!alpha)
-        {
-                ret = -1;
-                goto end;
-        }
-	/* TODO do rectangle, wise one color table should not be used for all rectangle */
-        mapclut_paletee(palette, alpha, (uint32_t *)rect[0].data[1],rect[0].nb_colors);
-	quantize_map(alpha, palette, pbuf, width*height, 3, rect[0].nb_colors);
 #ifdef ENABLE_OCR
-	str = ocr_bitmap(palette, alpha, pbuf, width, height, sub->lang_index);
-	if(str && str[0])
+	if (rect[0].ocr_text && *(rect[0].ocr_text))
 	{
 		if (context->prev_start != -1 || !(sub->flags & SUB_EOD_MARKER))
 		{
+			char *buf = (char *) context->buffer;
+			unsigned h1, m1, s1, ms1;
+			unsigned h2, m2, s2, ms2;
 			mstotime (ms_start,&h1,&m1,&s1,&ms1);
 			mstotime (ms_end-1,&h2,&m2,&s2,&ms2); // -1 To prevent overlapping with next line.
 			sprintf ((char *) context->buffer,"<p begin=\"%02u:%02u:%02u.%03u\" end=\"%02u:%02u:%02u.%03u\">\n",h1,m1,s1,ms1, h2,m2,s2,ms2);
-			write (context->out->fh, context->buffer,strlen(context->buffer) );
-			len = strlen(str);
-			write (context->out->fh, str, len);
+			write (context->out->fh, buf,strlen(buf) );
+			len = strlen(rect[0].ocr_text);
+			write (context->out->fh, rect[0].ocr_text, len);
 			write (context->out->fh, encoded_crlf, encoded_crlf_length);
-			sprintf ((char *) str,"</p>\n");
+			sprintf ( buf,"</p>\n");
+			write (context->out->fh, buf,strlen(buf) );
 
 		}
 	}
 #endif
 
-end:
 	sub->nb_data = 0;
 	freep(&sub->data);
-	freep(&palette);
-	freep(&alpha);
 	return ret;
 
 }
